Refactored FSM.ts into multiple files
fsmService in /services
other FSM related files in /FSM
